feat: add subjectId to fetchResource and move taxonomy stuff to separate file
